完成文章，评论，个人中心管理前后端对接

+++ hap-business/src/main/java/com/heartape/hap/business/entity/Label.java
@@ -26,7 +26,7 @@ public class Label extends BaseEntity {
     private Long labelId;
 
     @ApiModelProperty("最多3级")
-    private Long level;
+    private Integer level;
 
     @TableField(value = "parent_id",typeHandler = JacksonTypeHandler.class)
     private List<Long> parentId;
